Documented code and removed unnecessary code and files
