--- client/src/Components/ChangePasswordMenu/ChangePasswordMenu.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import './ChangePasswordMenu.css';
-
-const ChangePasswordMenu = () => {
-  return (
-    <div>ChangePasswordMenu</div>
-  )
-}
-
-export default ChangePasswordMenu
