+++ client/src/Components/DeleteFileMenu/DeleteFileMenu.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 import './DeleteFileMenu.css';
-import Popup from 'reactjs-popup';
 import 'reactjs-popup/dist/index.css';
 import { deleteFile } from '../../functions.js';
 
@@ -12,7 +11,6 @@ const DeleteFileMenu = ({ fileName, setSendDeleteMenu, setEditOrNewFile, userFil
     const handleFileDelete = async () => {
 
         const formData = {
-        //   email: localStorage.getItem('userEmail'),
           email: userEmail,
           fileName: fileName
         };
@@ -32,29 +30,27 @@ const DeleteFileMenu = ({ fileName, setSendDeleteMenu, setEditOrNewFile, userFil
       };
 
     return (
-        // <Popup trigger={<button>Trigger</button>} position="bottom">
-            <div className='deleteFileOuterMenuWrapper'>
-                <div className='deleteMenuInnerContainer'>
-                    <div className='deleteMenuHeaderContainer'>
-                        <h2>Are you sure you want to delete {fileName}?</h2>
-                    </div>
-                    <div className='deleteFileMenuButtonWrapper'>
-                        <button
-                            className='deleteFileMenuCancelButton'
-                            onClick={handleFileDelete}
-                        >
-                            Delete
-                        </button>
-                        <button 
-                            className='deleteFileMenuSendButton'
-                            onClick={() => setSendDeleteMenu(0)}
-                        >
-                            Cancel
-                        </button>
-                    </div>
+        <div className='deleteFileOuterMenuWrapper'>
+            <div className='deleteMenuInnerContainer'>
+                <div className='deleteMenuHeaderContainer'>
+                    <h2>Are you sure you want to delete {fileName}?</h2>
+                </div>
+                <div className='deleteFileMenuButtonWrapper'>
+                    <button
+                        className='deleteFileMenuCancelButton'
+                        onClick={handleFileDelete}
+                    >
+                        Delete
+                    </button>
+                    <button 
+                        className='deleteFileMenuSendButton'
+                        onClick={() => setSendDeleteMenu(0)}
+                    >
+                        Cancel
+                    </button>
                 </div>
             </div>
-        // </Popup>
+        </div>
     );
 }
