+++ client/src/Components/EditPage/EditPage.css
@@ -96,7 +96,6 @@
     display: flex;
     flex-direction: row;
     align-items: flex-start;
-    /* justify-content: space-around; */
     width: 100%;
 }
 
@@ -170,39 +169,3 @@
     border-style: solid;
     font-weight: 100;
 }
-
-/* .editScreenContainer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.homeFileSearchBar {
-    padding: 15px;
-    width: 600px;
-    border-radius: 50px;
-}
-
-.editScreenButtonWrapper {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    width: 100%;
-    margin-top: 60px;
-    margin-bottom: 100px;
-}
-
-.editScreenButton {
-    width: 18.75%;
-    padding: 15px;
-    background-color: orange;
-    color: white;
-    border-radius: 15px;
-    font-size: large;
-    cursor: pointer;
-}
-
-.editScreenImage {
-    object-fit: contain;
-    max-height: 1000px;
-}*/
