+++ client/src/Components/EditProfilePage/EditProfilePage.css
@@ -6,7 +6,6 @@
 
 .mainDiv {
     padding-top: 5px;
-    /* height: 100vh; */
     width: 100%;
     display: flex;
     flex-direction: row;
@@ -107,18 +106,6 @@
     width: 10px;
     background-color: darkblue;
     border-radius: 15px;
-}
-
-.changePasswordMenuOuterContainer {
-    /* position: fixed; */
-    /* width: 100%;
-    height: 100%;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: center; */
-    /* background-color: rgb(0, 0, 0, 0.5); */
-    /* top: 0; */
 }
 
 .changePasswordMenuInnerContainer {
