+++ client/src/Components/EditProfilePage/EditProfilePage.js
@@ -1,11 +1,10 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { updateAllData } from '../../reduxSlices/userSlice.js';
 import { useNavigate } from 'react-router-dom';
-import { useQuery } from 'react-query';
 import ArrowBackIcon from '@mui/icons-material/ArrowBack';
 import './EditProfilePage.css';
-import { editProfileInformation, editUserPassword, getUserInformation } from '../../functions.js';
+import { editProfileInformation, editUserPassword } from '../../functions.js';
 
 const EditProfilePage = () => {
 
@@ -14,19 +13,7 @@ const EditProfilePage = () => {
 
   const userData = useSelector((state) => state.userData);
 
-  // const { data, isLoading } = useQuery({
-  //   queryKey: "userData",
-  //   queryFn: () => getUserInformation(userData.email)
-  //   // queryFn: () => getUserInformation(localStorage.getItem("userEmail"))
-  // });
-
-  // useEffect(() => {
-  //   setUserFormData(data);
-  // }, [data]);
-
-  // const [userFormData, setUserFormData] = useState(data);
   const [userFormData, setUserFormData] = useState(userData);
-  // const [showChangePasswordMenu, setShowChangePasswordMenu] = useState(false);
   const [showChangeProfileError, setShowChangeProfileError] = useState(0);
   const [showChangePasswordError, setShowChangePasswordError] = useState(0);
   const [userPasswordData, setUserPasswordData] = useState({
@@ -35,19 +22,11 @@ const EditProfilePage = () => {
     reenterNewPassword: ""
   });
 
-  // useEffect(() => {
-  //   setUserFormData(userData);
-  // }, [userFormData, userData]);
-
-
   const navigateBack = () => {
     navigate('/home');
   };
 
   const handleUpdateData = async () => {
-
-    // userFormData.originalEmail = localStorage.getItem("userEmail");
-    // userFormData.originalEmail = userData.email;
 
     try {
 
@@ -63,11 +42,8 @@ const EditProfilePage = () => {
       const localUpdateDataResult = await editProfileInformation(localFormData);
             
       if (localUpdateDataResult.status === 200) {
-        // localStorage.setItem("userEmail", userFormData.email);
         dispatch(updateAllData(userFormData));
         setShowChangeProfileError(3);
-        // setShowChangeProfileError({ show: false, message: '' });
-        // navigate('/home');
       } else {
         setShowChangeProfileError(localUpdateDataResult.response.data);
       }
@@ -82,7 +58,6 @@ const EditProfilePage = () => {
     try {
 
       const localUserPasswordData = {
-        // email: localStorage.getItem("userEmail"),
         email: userData.email,
         oldPassword: userPasswordData.oldPassword,
         newPassword: userPasswordData.newPassword,
@@ -90,7 +65,6 @@ const EditProfilePage = () => {
       };
 
       const localChangePasswordResult = await editUserPassword(localUserPasswordData);
-      console.log("localChangePasswordResult: ", localChangePasswordResult);
       
       if (localChangePasswordResult.status === 200) {
         setUserPasswordData({
@@ -99,11 +73,8 @@ const EditProfilePage = () => {
           reenterNewPassword: ""
         });
         setShowChangePasswordError(4);
-        // setShowChangePasswordMenu(false);
-        // navigate('/home');
       } else {
         setShowChangePasswordError(localChangePasswordResult.response.data);
-        // setShowChangePasswordError();
       }
 
     } catch (err) {
@@ -119,7 +90,6 @@ const EditProfilePage = () => {
       </button>
       <hr style={{ width: '95%' }} />
       <div className='mainDiv'>
-        {/* {isLoading ? (<h1 style={{ color: 'white' }}>Loading...</h1>) : ( */}
           <div className='contentDiv'>
             <div className='editProfileHeaderWrapper'>
               <h1 className='editProfileHeaderText'>Edit Profile Information</h1>
@@ -132,8 +102,6 @@ const EditProfilePage = () => {
               <input 
                 className='itemInput' 
                 type='text'
-                // defaultValue={data?.firstname}
-                // defaultValue={userData.firstName}
                 value={userFormData.firstName} 
                 onChange={(e) => setUserFormData({ ...userFormData, firstName: e.target.value })} 
               />
@@ -152,12 +120,10 @@ const EditProfilePage = () => {
             <div className='itemDiv'>
               <div className='itemHeaderDiv'>
                 <h2 className='itemHeader'>Username</h2>
-                {/* { (showChangeProfileError === 2 || showChangeProfileError === 4) && <h6 style={{ color: 'red', marginLeft: '60px' }}>Username is already taken</h6> } */}
               </div>
               <input 
                 className='itemInput' 
                 type='text' 
-                // defaultValue={data?.username} 
                 value={userFormData.userName}
                 onChange={(e) => setUserFormData({ ...userFormData, userName: e.target.value })}  
               />
@@ -165,7 +131,6 @@ const EditProfilePage = () => {
             <div className='itemDiv'>
               <div className='itemHeaderDiv'>
                 <h2 className='itemHeader'>Email</h2>
-                {/* { (showChangeProfileError === 3 || showChangeProfileError === 4) && <h6 style={{ color: 'red', marginLeft: '60px' }}>Email is already taken</h6> } */}
                 { showChangeProfileError === 1 && <h6 style={{ color: 'red', marginLeft: '60px' }}>Email is already in use</h6> }
                 { showChangeProfileError === 2 && <h6 style={{ color: 'red', marginLeft: '60px' }}>Username is already in use</h6> }
                 { showChangeProfileError === 3 && <h6 style={{ color: 'green', marginLeft: '60px' }}>Sucessfully updated user's information</h6> }
@@ -173,18 +138,14 @@ const EditProfilePage = () => {
               <input
                 className='itemInput' 
                 type='text' 
-                // defaultValue={data?.email}
                 value={userFormData.email} 
                 onChange={(e) => setUserFormData({ ...userFormData, email: e.target.value })}
               />
             </div>
             <div className='editButtonWrapper'>
-              {/* <button className='editMenuButton' onClick={navigateBack}>Back</button> */}
-              {/* <button className='editMenuButton' onClick={() => setShowChangePasswordMenu(true)}>Change Password</button> */}
               <button className='editMenuButton' onClick={handleUpdateData}>Save</button>
             </div>
           </div>
-        {/* )} */}
         <div className='verticalLineContainer'>
 
         </div>
@@ -232,7 +193,6 @@ const EditProfilePage = () => {
             { showChangePasswordError === 3 && <h6 className='changePasswordErrorText'>New password cannot be the same as the old one</h6> }
             { showChangePasswordError === 4 && <h6 className='changePasswordSuccessText'>Sucessfully updated password</h6> }
             <div className='changePasswordButtonWrapper'>
-              {/* <button className='changePasswordButton' onClick={navigateBack}>Back</button> */}
               <button className='changePasswordButton' onClick={handleChangePassword}>Save</button>
             </div>
           </div>
