+++ client/src/Components/FileCard/FileCard.css
@@ -9,17 +9,12 @@
 }
 
 .fileCardDiv:hover {
-    /* opacity: 0.5; */
     background: rgba(0,0,0,.5);
     cursor: pointer;
 }
 
 .fileCardDiv:active {
     background: rgba(0,0,0,.25);
-}
-
-.fileNameH1 {
-    /* color: white; */
 }
 
 .fileNameH2 {
