+++ client/src/Components/FileDescription/FileDescription.css
@@ -1,5 +1,4 @@
 .fileDescriptionListButton {
-    /* background-color: rgb(0, 0, 0, 0); */
     border-radius: 30px;
     border-color: rgb(0, 0, 0, 0);
     padding-left: 10px;
