--- client/src/Components/FileEditor/FileEditor.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import './FileEditor.css';
-
-const FileEditor = () => {
-  return (
-    <div>FileEditor</div>
-  )
-}
-
-export default FileEditor
