+++ client/src/Components/FileListDisplay/FileListDisplay.css
@@ -1,10 +1,8 @@
 .fileListDisplayMainWrapper {
     width: 100%;
-    /* height: 300px; */
     border: solid;
     border-width: 2px;
     border-color: #e8e8e8;
-    /* border-radius: 25px; */
 }
 
 .fileListDisplayHeaderWrapper {
@@ -42,7 +40,6 @@
     flex-direction: column;
     align-items: center;
     overflow-y: scroll;
-    /* height: 190px; */
 }
 
 .fileListDisplayFileGridDiv::-webkit-scrollbar {
@@ -77,9 +74,7 @@
     height: 30px;
     border: none;
     border-radius: 15px;
-    /* background-color: darkorange; */
     background-color: darkblue;
-    /* border: none; */
     color: white;
     font-size: large;
     cursor: pointer;
