--- client/src/Components/Home/Home.css
@@ -1,217 +1,0 @@
-.fileGridDiv {
-    width: 75%;
-    margin-left: 12.5%;
-}
-
-.homeFileSearchBar {
-    padding: 15px;
-    width: 600px;
-    /* margin-left: 30px; */
-    border-radius: 50px;
-}
-
-.homeSearchButton {
-    padding: 15px;
-    border-radius: 25px;
-    background-color: orange;
-}
-
-.searchWrapper {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-top: 50px;
-}
-
-.topSearchWrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: '50%';
-}
-
-.addButtonWrapper {
-    margin-top: 10px;
-    margin-bottom: 25px;
-}
-
-.addButton {
-    width: 300px;
-    padding: 15px;
-    border-radius: 15px;
-    background-color: darkorange;
-    color: white;
-    font-size: large;
-}
-
-.fileListButton {
-    background-color: rgb(0, 0, 0, 0);
-    border-color: rgb(0, 0, 0, 0);
-}
-
-.editScreenContainer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.editScreenButton {
-    width: 18.75%;
-    padding: 15px;
-    background-color: orange;
-    color: white;
-    border-radius: 15px;
-    font-size: large;
-}
-
-.editScreenButtonWrapper {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    width: 100%;
-    margin-top: 60px;
-}
-
-/* .uploadFileInput {
-    display: none;
-} */
-
-/* .uploadFileButton {
-    display: inline-block;
-    text-align: center;
-    width: 300px;
-    padding: 15px;
-    border-radius: 15px;
-    background-color: darkorange;
-    color: white;
-    font-size: large;
-} */
-
-.fileEditor {
-    background-color: 'white';
-    width: '75%';
-    height: '650px';
-}
-
-.sendFileOuterMenuWrapper {
-    width: 100%;
-    height: 100%;
-    background-color: rgb(0, 0, 0, 0.5);
-    z-index: 3;
-    position: absolute;
-    top: 0;
-}
-
-.sendFileInnerMenuWrapper {
-    height: 50%;
-    width: 75%;
-    margin-top: 25%;
-    margin-left: 12.5%;
-    background-color: white;
-    border-radius: 30px;
-}
-
-.sendFileMenuSearchBar {
-    padding: 15px;
-    width: 600px;
-    margin-left: 30px;
-    margin-top: 30px;
-    border-radius: 50px;
-}
-
-.userSearchSuggestionListBox {
-    padding: 15px;
-    background-color: white;
-    border-color: black;
-    border-style: solid;
-    border-width: thin;
-    border-radius: 30px
-}
-
-.userSearchSuggestionListButton {
-    background-color: white;
-    border: none;
-    font-size: large;
-    width: 100%;
-    text-align: left;
-    cursor: pointer;
-}
-
-.userSearchSuggestionListButton:active {
-    opacity: 0.5;
-}
-
-.userSearchSuggestionListWrapper {
-    width: 600px;
-    margin-left: 30px;
-    height: 50px;
-    position: absolute;
-}
-
-.sendFileMenuSendButton {
-    height: 150px;
-    width: 35%;
-    background-color: blue;
-    border-radius: 30px;
-    color: white;
-    font-size: xx-large;
-    cursor: pointer;
-}
-
-.sendFileMenuSendButton:hover {
-    opacity: 0.5;
-}
-
-.sendFileMenuSendButton:active {
-    opacity: 0.9;
-}
-
-.sendFileMenuCancelButton {
-    height: 150px;
-    width: 35%;
-    background-color: red;
-    border-radius: 30px;
-    color: white;
-    font-size: xx-large;
-    cursor: pointer;
-}
-
-.sendFileMenuCancelButton:hover {
-    opacity: 0.5;
-}
-
-.sendFileMenuCancelButton:active {
-    opacity: 0.9;
-}
-
-.sendFileMenuButtonWrapper {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-    margin-top: 75px;
-}
-
-.deleteMenuInnerContainer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    height: 50%;
-    width: 75%;
-    margin-top: 25%;
-    margin-left: 12.5%;
-    background-color: white;
-    border-radius: 30px;
-}
-
-.deleteFileMenuButtonWrapper {
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-    margin-top: 75px;
-}
-
-.deleteMenuHeaderContainer {
-    margin-top: 50px;
-}
