+++ client/src/Components/HomePage/HomePage.css
@@ -28,48 +28,9 @@
 .homeFileSearchBar {
     padding: 15px;
     width: 600px;
-    /* margin-left: 30px; */
     margin-top: 50px;
     border-radius: 50px;
 }
-
-/* .addButtonWrapper {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-}
-
-/* .addButtonWrapper {
-    margin-top: 10px;
-    margin-bottom: 25px;
-} */
-
-/*.addButton {
-    width: 110px;
-    height: 30px;
-    border: none;
-    border-radius: 15px;
-    background-color: darkblue;
-    color: white;
-    font-size: large;
-    cursor: pointer;
-} */
-
-/* .uploadFileButton {
-    display: inline-block;
-    text-align: center;
-    width: 110px;
-    height: 30px;
-    border-radius: 15px;
-    background-color: darkblue;
-    cursor: pointer;
-    color: white;
-    font-size: large;
-}
-
-.uploadFileInput {
-    display: none;
-} */
 
 .fileGridDiv {
     width: 100%;
@@ -91,27 +52,6 @@
 .fileGridDiv::-webkit-scrollbar-thumb {
     background-color: grey;
 }
-
-/* .inboxRecycleFileGridDiv {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    overflow-y: scroll;
-    height: 190px;
-}
-
-.inboxRecycleFileGridDiv::-webkit-scrollbar {
-    width: 5px;
-}
-
-.inboxRecycleFileGridDiv::-webkit-scrollbar-track {
-    background-color: #e8e8e8;
-}
-
-.inboxRecycleFileGridDiv::-webkit-scrollbar-thumb {
-    background-color: grey;
-} */
 
 .fileListButton {
     background-color: rgb(0, 0, 0, 0);
@@ -147,35 +87,6 @@
     opacity: 0.5;
 }
 
-/* .inboxWrapper {
-    width: 100%;
-    height: 300px;
-    border: solid;
-    border-width: 2px;
-    border-color: #e8e8e8;
-    border-radius: 25px;
-} */
-
-/* .inboxHeaderWrapper {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-evenly;
-} */
-
-/* .inboxIconTextWrapper {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-} */
-
-/* .inboxHorizontalLine {
-    border: none;
-    background-color: #e8e8e8;
-    height: 2px;
-    width: 95%;
-} */
-
 .recycleBinWrapper {
     width: 100%;
     height: 300px;
@@ -194,8 +105,6 @@
 
 .myFilesWrapper {
     width: 45%;
-    /* border: solid; */
-    /* border-width: 2px; */
     border-color: #e8e8e8;
     border-radius: 25px;
 }
@@ -206,20 +115,6 @@
     align-items: center;
     justify-content: space-evenly;
 }
-
-/* .myFilesSearchWrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-} */
-
-/* .myFilesSearchWrapper {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: center;
-} */
 
 .myFilesButtonWrapper {
     display: flex;
@@ -266,17 +161,6 @@
 .myFilesDropDownMenuButton:hover {
     background-color: rgb(0, 0, 0, 0.25);
 }
-
-/* .uploadFileButtonDropDown {
-    padding: 15px;
-    cursor: pointer;
-    border-radius: 15px;
-    text-align: center;
-}
-
-.uploadFileButtonDropDown:hover {
-    background-color: rgb(0, 0, 0, 0.25);
-} */
 
 .uploadFileInputDropDown {
     display: none;
