+++ client/src/Components/Navbar/Navbar.css
@@ -4,7 +4,6 @@
     justify-content: space-between;
     align-items: center;
     background-color: darkblue;
-    /* background-image: linear-gradient(90deg, #ff6600 0%, #ff6600 35%, #ff853e 100%); */
     border-bottom-left-radius: 10px;
     border-bottom-right-radius: 10px;
     position: sticky;
@@ -28,7 +27,6 @@
 }
 
 .navBarMenuButton {
-    /* background-color: #ff853e; */
     background-color: darkblue;
     border: thin;
     border-color: white;
