+++ client/src/Components/Navbar/Navbar.js
@@ -1,6 +1,5 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { useSelector } from 'react-redux';
-import { useNavigate } from 'react-router-dom';
 import FolderIcon from '@mui/icons-material/FolderOutlined';
 import './Navbar.css';
 import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown';
@@ -8,19 +7,9 @@ import NavbarMenu from '../NavbarMenu/NavbarMenu';
 
 const Navbar = () => {
 
-  const navigate = useNavigate();
   const isLoggedIn = useSelector((state) => state.userData.isLoggedIn);
 
   const [showMenu, setShowMenu] = useState(false);
-  // const [isLoggedIn, setIsLoggedIn] = useState(false);
-
-  // useEffect(() => {
-
-  //   if (localStorage.getItem("isLoggedIn")) {
-  //     setIsLoggedIn(true);
-  //   }
-
-  // });
 
   const handleNavbarMenuClick = () => {
 
@@ -44,7 +33,7 @@ const Navbar = () => {
             <ArrowDropDownIcon style={{ color: 'white' }} />
           </button>
           {showMenu ? (
-            <NavbarMenu /*setIsLoggedIn={setIsLoggedIn}*/ setShowMenu={setShowMenu} />
+            <NavbarMenu setShowMenu={setShowMenu} />
           ) : (<></>)}
         </div>
         ) : (<></>)}
