+++ client/src/Components/NavbarMenu/NavbarMenu.js
@@ -4,16 +4,14 @@ import { updateLoggedInStatus, resetUserData } from '../../reduxSlices/userSlice
 import { useNavigate } from 'react-router-dom';
 import './NavbarMenu.css';
 
-const NavbarMenu = ({ /*setIsLoggedIn,*/ setShowMenu }) => {
+const NavbarMenu = ({ setShowMenu }) => {
   
   const navigate = useNavigate();
   const dispatch = useDispatch();
 
   const handleLogout = () => {
-    // localStorage.clear();
     dispatch(resetUserData());
     setShowMenu(false);
-    // setIsLoggedIn(false);
     dispatch(updateLoggedInStatus(false));
     navigate('/');
   }
@@ -22,10 +20,6 @@ const NavbarMenu = ({ /*setIsLoggedIn,*/ setShowMenu }) => {
     navigate('/editprofile');
     setShowMenu(false);
   }
-
-  const handleChangePassword = () => {
-    setShowMenu(false);
-  };
 
   return (
     <div className='navbarMenuWrapper'>
