+++ client/src/Components/SearchBar/SearchBar.css
@@ -1,7 +1,6 @@
 .searchBar {
     padding: 12px;
     border-width: thin;
-    /* width: 300px; */
     width: 100%;
 }
