--- client/src/reduxSlices/counter/counterSlice.js
@@ -1,23 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit';
-
-export const counterSlice = createSlice({
-    name: 'counter',
-    initialState: {
-        value: 0,
-    },
-    reducers: {
-        increment: (state) => {
-            state.value += 1
-        },
-        decrement: (state) => {
-            state.value += 1
-        },
-        incrementByAmount: (state, action) => {
-            state.value += action.payload
-        }
-    }
-});
-
-export const { increment, decrement, incrementByAmount } = counterSlice.actions
-
-export default counterSlice.reducer
